server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.116.1",
    "prisma>=0.15.0",
    "pydantic>=2.11.7",
    "python-jose>=3.5.0",
]
